Revert "refactor: Refactor god class - introduce partition and queue state classes"
This reverts commit f02b5e0cce01838d9a6758d0f77e856cb5f062e8.

+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/DrainingCloseable.java
@@ -1,8 +1,9 @@
 package io.confluent.parallelconsumer;
 
 /*-
- * Copyright (C) 2020-2021 Confluent, Inc.
+ * Copyright (C) 2020 Confluent, Inc.
  */
+
 import java.io.Closeable;
 import java.time.Duration;
 
@@ -73,6 +74,6 @@ public interface DrainingCloseable extends Closeable {
      *
      * @return the number of consumed but outstanding records to process
      */
-    long workRemaining();
+    int workRemaining();
 
 }
